+++ core/integration/setup_test.go
@@ -59,7 +59,10 @@ func (t tstReporter) Fatalf(format string, args ...interface{}) {
 	os.Exit(1)
 }
 
-var marketConfig = market.NewMarketConfig()
+var (
+	marketConfig    = market.NewMarketConfig()
+	successorConfig = market.NewSuccessorConfig()
+)
 
 type executionTestSetup struct {
 	cfg              execution.Config