+++ core/types/governance_proposal.go
@@ -101,6 +101,8 @@ const (
 	ProposalErrorLinearSlippageOutOfRange ProposalError = vegapb.ProposalError_PROPOSAL_ERROR_LINEAR_SLIPPAGE_FACTOR_OUT_OF_RANGE
 	// ProposalErrorSquaredSlippageOutOfRange squared slippage factor is negative or too large.
 	ProposalErrorQuadraticSlippageOutOfRange ProposalError = vegapb.ProposalError_PROPOSAL_ERROR_QUADRATIC_SLIPPAGE_FACTOR_OUT_OF_RANGE
+	// ProposalErrorInvalidSuccessorMarket indicates the successor market parameters were invalid.
+	ProposalErrorInvalidSuccessorMarket ProposalError = vegapb.ProposalError_PROPOSAL_ERROR_INVALID_SUCCESSOR_MARKET
 )
 
 type ProposalState = vegapb.Proposal_State
@@ -218,6 +220,29 @@ func (p *Proposal) MarketUpdate() *UpdateMarket {
 	default:
 		return nil
 	}
+}
+
+func (p *Proposal) IsNewMarket() bool {
+	return p.Terms.Change.GetTermType() == ProposalTermsTypeNewMarket
+}
+
+func (p *Proposal) NewMarket() *NewMarket {
+	switch terms := p.Terms.Change.(type) {
+	case *ProposalTermsNewMarket:
+		return terms.NewMarket
+	default:
+		return nil
+	}
+}
+
+func (p *Proposal) IsSuccessorMarket() bool {
+	if p.Terms == nil || p.Terms.Change == nil {
+		return false
+	}
+	if nm := p.NewMarket(); nm != nil {
+		return nm.Changes.Successor != nil
+	}
+	return false
 }
 
 func (p *Proposal) WaitForNodeVote() {